rename schema_model module to model
