+++ pandera/model.py
